CC-1845 Re-implement concept browse tree transform
Re-implement the concept browse tree transform.

The ConceptTreeTransformProvider is a replacement for
JsonTreeTransformProvider.

There are changes to the Registry Schema and associated XML Schema
files. There is a new browse-flag element contained within the version
element.

See MigrateJsonTreesToConceptTrees for a migration tool, which can be
run using the Ant task migrate-jsontree-artefacts.

Note: corresponding changes must be made to the Portal!