CC-1845 Don't always output value for defaultSortByNotation
A value for defaultSortByNotation was always being put into the
generated concepts_tree2.json file, even if maySortByNotation is
false. It's better not to include a value for the
defaultSortByNotation flag in that case.

So don't include a (default false) value for defaultSortByNotation in
the generated concepts_tree2.json file, if maySortByNotation is false.