CC-2541 CC-2542 Fix lookupIRI's JPA query to use entity IDs
The JPA query underlying lookupIRI is broken: it twice uses an "id"
attribute (i.e., a surrogate key) when it should use the appropriate
entity ID.

Fix the test data so that the defect is picked up.

Then, fix the query to use the entity IDs instead of the surrogate
IDs.

As reported in SD-213325.