CC-2543 Don't fail to create Solr doc if Jackson barfs on concept list
It seems it's possible to construct RDF data that causes Jackson to
fail to parse the CONCEPT_LIST version artefact during creation of the
Solr document.

For now, if that happens, log the problem but don't fail. The concepts
won't then be included in the Solr document.

Can we do better? Revisit later with a better solution.