CC-2570 Add "sort" parameter to search API method
For the search API method, support a "sort" parameter, which
influences the sort order of search results.

Define the possible values as an enumerated type in the Registry
Schema.

The values of the enumerated type are (for now) "aToZ", "zToA", and
"relevance".

Use a "sensible" default sort order when one is not specified, which
is consistent with the previous behaviour:

* When no search query term is specified, default to "aToZ".
* When a search query term is specified, default to "relevance".