CC-2588 Start to implement version 2 of Solr collection scheme
Remove the code to install version 1 of the scheme.

Start to implement version 2 of the Solr collection scheme.  Define a
new field type for "Google-like" phrase searches. Add a couple of new
fields that use the new field type.

Fix the existing defect that the publisher_search field was not being
copied into.

For version 2 of the scheme, load the custom solrconfig.xml.
(This now requires ZooKeeper.)

Explain the additional required Solr configuration properties in
registry.properties.sample.

Update build.xml's create-solr-schema target to pass in the additional
parameters required by CreateSchema.