Fix Ant build dependency for generate-jaxb-source target
The generate-jaxb-source Ant target always touched the timestamp file,
and that means, for example, that you can't run targets such as
create-solr-schema unless you are running as the owner of the files
... and we want to be able to run "ant create-solr-schema" as a
different user.

It seems this is a long-standing defect.

In commit e165ca441d59ed1bb2ab550342f45193f4a83648 I introduced the
check-jaxb-source target, specifically to make use of the generated
timestamp file, but I then failed to add check-jaxb-source as a
dependency of the generate-jaxb-source target.

The check-jaxb-source target is also broken, as the wildcards of the
include elements were wrong: they didn't match any of the files they
were supposed to match!

So:

* Fix the check-jaxb-source target to use the correct wildcards.
* Add check-jaxb-source as a dependency of the generate-jaxb-source
  target.

And while we're here:

* Tidy to use the ${src-jaxb-gen} property in two places, and fix an
  indentation error.