Add Safari Press Solr query plugin and start to use it
Add the JAR and source for the Safari Press Solr query plugin.

In SearchIndex, now add the pqf field to queries, and add the
corresponding phrase fields to the highlighting fields.

Note: we haven't yet configured loading of the plugin in
solrconfig.xml, so we're not getting results from it yet ... that's
the next step.