Use POST method to send queries to Solr
The text of Solr queries can get long, so force the use of the POST
method when sending a query to Solr. (If the query exceeds 8192
characters, the default GET method fails.)

See https://issues.apache.org/jira/browse/SOLR-13014

Sigh: this is currently _broken_ for EmbeddedSolrServer, so for the
test suite, we need to keep using the default (GET) method for now.

See https://issues.apache.org/jira/browse/SOLR-12858