Solr schema: copy top_concept to concept_search, not subject_search
The original schema definition from May 2015 had this:

 <copyField source="top_concept" dest="subject_search"/>

Why? It's not clear. It may be a copy/paste error from the previous
line.  The top_concept field should be copied to a field
concept_search. But that field was only created in August 2015.

The original definition works to give search results, but it gives
"wrong" (because misleading) _highlighting_ information.

We copied the original behaviour into the Registry.

Since we're taking time to pay attention to highlighting information,
fix it now.