Unescape HTML entities when creating Solr document
When creating a Solr document, be more careful when working with HTML
content that contains entity references. It's not good to include
something like "A &ldquo; B" in the Solr document, as that tends to
get indexed as "A ldquo B", which breaks phrase searches.

So apply unescaping to entities in HTML content. That turns an entity
reference such as &ldquo; into the corresponding Unicode
character. Then it can be dealt with by indexing as desired, and the
stored value is still OK to return in a search result.