Be non-specific in comment about the values of hl.simple.{pre,post}
In commit 30d8b6c8b5b805c86e5fefd87e429cca85f80cc4, I changed the
hl.simple.pre and hl.simple.post parameters, but omitted to change a
comment that specified the significance of the old values.

Update that comment to be non-specific about the value of the
parameters.