CC-2619 Apply filtering as a conjunction of disjunctions per facet
RVA search results have always been a conjunction of the query term
with _all_ filters.

For example, if the query term is Q, and filtering is being applied
with two subjects S1 and S2, and two publisher values P1 and P2, then
the query is executed as:

Q AND subject = S1 AND subject = S2
  AND publisher = P1 AND publisher = P2.

This is wrong.

The query term should be conjoined with a _disjunction_ for each
facet.

The example query should be executed as:

Q AND (subject = S1 OR subject = S2)
  AND (publisher = P1 OR publisher = P2).

Change the construction of the Solr query to match the correct
behaviour.