CC-2615 Index non-ISO-639 language tags without lowercasing them
Inside the Registry database, language values are "supposed to be" ISO
639 tags, e.g., "en", "de". The Portal offers a selection of
"human-readable" values, e.g., "English", "German", and translates
into/from the ISO 639 tags as needed.

But there's nothing stopping an API user from providing a value that's
not ISO 639, e.g., "Turkish". I think that's not a problem per se, at
least for now.

We resolve tags using the ICU4J library when we create the Solr
document for indexing. When we do that, "en" becomes "English", but
"Turkish" becomes "turkish".

Check the result of language tag resolution, and make sure that
"Turkish" stays as "Turkish". In general, let's say: apply resolution
using ICU4J, and if the result is case-insensitively the same as the
value pre-resolution, use the pre-resolution value as the result of
the resolution instead.

Refactor language resolution into a separate method, and rewrite the
tests in TestLanguageConversion to use that method.