Refactor Solr schema creation class
We're going to add a second Solr collection, in which we'll store
information about resources (e.g., SKOS Concepts).

In preparation for that, refactor CreateSchema into an abstract "base"
class SolrSchemaBase with the utility methods, and a child class
CreateSolrSchemaRegistry that contains the code specific to the Solr
registry schema.