CC-2619 Specify JSON facets using "filter" and "excludeTags"
As part of Solr requests, we now ask for facet information for facet
values/counts for values not currently in search results.  But we were
asking in a wrong way. We use the JSON facet API, but we specified a
domain using "query", which always disregards a top-level query
term.

The result was that the "extra" facet counts were being computed
without regard to any specified query term.

Specify the JSON facets using a combination of "filter" and
"excludeTags" instead.