CC-2619 Specify JSON facets in a more efficient way
Further to commit d5e0aa271bd9d94f5b128a5dadf89332a36409f0, redo the
specification of the JSON facets in a more efficient way.

Since we're using "filter", for each facet f, rather than specifying
for f an excludeTags setting that includes all facets and a filter
that includes the top-level filters for all other facets, do neither
instead!

For facet f, just specify excludeTags with f, and a filter value that
is the negation of the top-level filter values for f.