Index all vocabs: delete existing docs only when new ones are ready
In testing with indexing a very large number (> 9000) of vocabularies,
creating the Solr documents for all vocabularies can take over a
minute. It used to be that AdminRestMethods.indexAll() invoked first
unindexAllVocabularies() and then indexAllVocabularies(). That meant
that you (i.e., a portal or API user) could be without an index for
over a minute. That would be bad.  So I've copied the call to
SOLR_CLIENT.deleteByQuery() into indexAllVocabularies(), and (for now)
we no longer need the unindexAllVocabularies() method.  If, in future,
we decide to add an AdminRestMethods.unindexAll() method, we can
restore that method.