CC-2612 When logging non-literal object, also log the subject
When ResourceDocsTransformProvider logs that a "predicate of interest"
has a non-literal object, log not just the prediate and object, but
also log the subject.

Note for the record: according to the SKOS Reference, it's not at all
an error to give a non-literal object. But we can't do much with an
IRI as the object. Well, at least, for _now_ we're not going to do
anything with it.