Remove no-longer-in-play code to set up Solr "raw JSON" responses
Since commit 78fb030b77dd50eb537812b7afed183c1f9d71d6, we no longer
request the "raw JSON" response format from Solr, and we use our own
custom JSON writer to send back responses to API clients. But the code
to wrap the Solr query in such a request was erroneously left
behind. It was doing no harm, but it wasn't doing _anything_ either,
so it's confusing to leave it there.

Remove the no-longer-in-play code that set up the "raw JSON" request.