Support nested SolrDocumentLists in Solr responses
Since 78fb030b77dd50eb537812b7afed183c1f9d71d6, we use a custom JSON
writer to create responses to API search requests.  We do this in
order to mimic the nesting structure of JSON (i.e., the way that
objects and arrays are used) in Solr's own responses.

It turns out that the mimicking fails when there are SolrDocumentLists
nested at lower levels, such as inside the "expanded" component of
responses (i.e., when the collapsing query parser and the parameter
"expand=true" are used). Such SolrDocumentLists are being written as
"normal" arrays, so that numFound, etc., values are getting lost.

So redo the custom JSON writer to cope with nested SolrDocumentLists,
so that the responses mimic Solr's.