CC-2612 Resource search: trim RDF types not of interest
We only index resources that have an RDF type "of interest" (e.g.,
SKOS Concepts). But we index _all_ RDF types of such resources. That
means that we have resources in our index that _also_ have RDF types
that are _not_ of interest, e.g., rdfs:Resource.

Now, trim RDF types "not of interest" for the Solr documents, so that
the Solr documents only have values for the rdf_type field that are
"of interest".