CC-2612 Resource search: expand expanded results and group
In the expanded results for an IRI, include _all_ instances of the
IRI, not just ones for which the Solr document matches the top-level
query and filters.

Include more results in the expanded results for each IRI.

Within the expanded results for each IRI, sort by vocabulary ID (i.e.,
to group results for the same vocabulary ID together), then give the
"most-recently-released" (as far as we can tell) version first.