CC-2612 Resource search: change version_id field to pint type
In the expanded results for an IRI, we want to group by vocabulary_id,
then sort in reverse order of release_date. But if there are two
versions with the same version_id, we should sort in reverse order of
version_id.

We had made version_id a string field, which means values are sorted
as strings, not numerically. That means, e.g., "100" < "99".

Since we don't need version_id to be a string field -- it's not a
field of the Solr registry collection -- we can change its type to
pint, and then its values will be sorted numerically.

So change the type of the version_id field to pint.