CC-2612 Resource search: run ResourceDocs transform after publish
Oops, I had set the priority of the ResourceDocs transform to be the
same as the other concept-related transforms, i.e., before import.

But it has to be run _after publish_, as we only know after publish if
there is a SISSVOC access point for us to use as the value of the
sissvoc_endpoint field of generated Solr documents.

So change the priority of the ResourceDocs transform to be after
publish for insert/perform, and "after" (i.e., before!) publish for
delete.

Adjust test data accordingly.