CC-2612 Resource search: analytics for lang, support _all as value
As part of constructing the logstash log entry, take note of the
languages only _after_ we are sure what they are. I.e., handle the
absence of a value for the "language" filter.

Support passing in "_all" as a value for the "language" filter. That
value was previously being converted to "-_all" before being appended
to field names.

Take the opportunity to add a clarifying comment explaining why we
have to have both the _search and _phrase fields in the highlight
fields.