CC-2612 Resource search: don't index empty data; prefer prefLabels
1. Don't index empty/blank string values in the resources collection.

2. When selecting a title for a resource (i.e.. a value for the
"title" field), we look at SKOS prefLabels and altLabels.

We used to look for a prefLabel and then an altLabel for the primary
language, then for a prefLabel and then an altLabel for no language,
then for a prefLabel and then an altLabel for the "en" language tag.

Now, look for the three possible prefLabels before considering looking
for any altLabels.