CC-2612 Resource search: add vocab/ver-level fields at index time
The Solr document for a resource includes vocabulary- and
version-level metadata. If the vocabulary-level metadata is updated,
we would need to re-run the transform for every version!  And this
would indeed need to be done every time we run updateVocabulary, as
one of the fields of the Solr document is the vocabulary last_updated
value.

This is bad.

So instead, when creating the Resource Docs version artefact, _don't_
insert values for the fields containing the vocabulary- and
version-level metadata, e.g., vocabulary title and owner, and version
title. Rather, get the values at _indexing_ time, and insert them into
the Solr documents before sending them to Solr.