RVA-5 Resource search: put a limit on pp value
There's no limit on the number of results that can be requested from
either of the search API methods.

In the case of vocabulary search, we can live with that, because there
are (in production and demo) only a few hundred results.

But now we have the resource search, and there are getting up to
100000 Solr documents.

Let's add a "sensible" limit to the number of results that can be
requested through the resource search API method.

For now, limit the number of results to 500.

See also CC-2681.