Analytics: for GeoIP, if host is localhost, do lookup by public IP
The GeoIP lookup doesn't work on any of the IP addresses for
localhost (i.e., either 127.0.0.1 or ::1).

So, when looking up the geolocation of the caller's IP address, if the
IP address is for localhost (either IPv4 or IPv6), then get the
server's publicly-visible IP address and look up _that_ value instead.