Notification CollectEvents: log unknown owner name at warn level
The notification CollectEvents class adds a log entry when it looks
for an owner's "full name" but doesn't find it.  We have been adding
these log entries at "error" level.

But in fact this is not necessarily an "error", as we have our "ARDC
Curated" role which we deliberately _don't_ put in the roles database.

So log at "warn" level instead.