Limit size of debugging output generated from incoming XML
The validation of incoming, user-supplied vocabulary and related
entity metadata includes logging parsed content at debug level.

However, if provided in XML format using entities, the result can be
very long, and we don't want to log "too much" of it, especially if it
is in fact part of a DoS attack!

So introduce a utility class LimitedSizeStringWriter, which is a
StringWriter that puts a maximum limit on the content that can be
written. Use that in the places in vocabulary/related entity
validation where a StringWriter is used.