Simply search implementation, since filtersJson is always non-null
The query() method of both the SearchRegistryIndex and
SearchResourcesIndex classes has different behaviour based on a test
to see if the value of the filtersJson parameter is non-null.

But in fact, filtersJson is guaranteed to be non-null, because the
calling API methods have a @NonNull annotation on the incoming
parameter.

So change the query() method simply to _require_ that filtersJson be
non-null, and refactor accordingly.