Include Hibernate Validator config files in the WAR file
Commit 4a939016b54b7642f472b20f983ce33cb7c2bd73 of March 2017
included this note in the comment:

"Provide basic Hibernate Validator configuration that uses
ReflectionParameterNameProvider to get prettier names for parameters
in constraint violations."

But I failed to include these config files in the generated WAR file!

At last, include the config files in the generated WAR file to get
prettier names for parameters in constraint violations.