RVA-4 Resource search: log ids/owners of expanded results
For resource search events, log the Solr document ids and owners of
expanded results. Unlike the ids/owners of the "main" results, which
are two ordered lists of the same length whose values correspond, the
values logged have duplicates removed, and so the two lists may have
different lengths, and the values can't be matched across the two
lists.

See also CC-2646.