Search: add count_only filter to get only numFound
The Portal requests two searches for every change in filters: a
registry and a resource search. But because of the Portal's tabbed
interface, the Portal user can only see one set of results at a
time. So support a "count_only" filter that only gives the numFound
value, but no search results (or facet counts or highlighting, etc.).