RVA-4 Change type of result_ids for registry search to string
Commit 8032e6964aabb56460f8b2bc85fa92c35875631c changed the logstash
logging for resource searches so that the type of
result_ids was no longer an integer but now a string.

This is problematic, because the registry search has a field of the
same name which is still an integer type.

So now, harmonize the registry search's field type to match.

I.e., for resource search events, log the Solr document ids in the
result_ids field with type string.