Compile test classes with -parameter for param names for validation
Commit d9414cbeed326cce1d737529356dec3321431ee1 included the
Hibernate Validator config files in the WAR file. But I forgot to
modify the test cases that use it. And it turns out that I had
also failed to add the -parameter for javac when compiling test
classes. (The test cases make use of the
RegistrySchemaValidationHelper class, and we need that class to
be compiled with -parameter, in order to get parameter names by
reflection.)

So, add -parameter to the javac command-line arguments when
compiling test classes. And update the validation test cases to
make use of the parameter names.