Fetch metadata from PoolParty: get all languages, strip suffixes
The API method /api/services/PoolParty/{serverId}/{projectId}/metadata
returns data with keys derived from the filenames in which the
metadata is stored, e.g., "concepts.trig", "adms.trig", etc..

Now, strip the suffix component of the filenames, so that the keys are
"concepts", "adms", etc.  This helps make the Portal code simpler.

And now also include a predicate key "ppcl:availablelanguages" that
contains the language tags for all the project's languages.