Fetch metadata from PoolParty: strip suffix from filename keys
The API method /api/services/PoolParty/{serverId}/{projectId}/metadata
returns data with keys derived from the filenames in which the
metadata is stored, e.g., "concepts.trig", "adms.trig", etc..

Now, strip the suffix component of the filenames, so that the keys are
"concepts", "adms", etc.  This helps make the Portal code simpler.