RVA-23 When deleting a file AP (but not the version), run subtasks
Suppose we have a version with two FILE access points.

We edit the version, and delete one of them.

We expect the concept transforms (JsonList, ConceptTree, and the new
ResourceDocs) to be run in this case. But they aren't. (In fact, you
don't need two FILE access points to start with; you could have a
WEB_PAGE AP and a FILE AP, and you delete just the FILE AP.)

The WorkflowMethods.deleteAccessPoint() method current makes the
caller choose between _either_ deleting/making historical the AP, _or_
running subtasks, but not both.

So modify WorkflowMethods.deleteAccessPoint() so that the method's
return value is a Pair, of which the two contained elements are:
* a Boolean, indicating whether or not the caller should delete/make
  historical the AP, and
* a (possibly null) list of subtasks to be performed.

This makes the two issues orthogonal: now the caller may or may not
delete/make historical the AP, and may or may not have subtasks to be
performed.

Modify the same method so that it takes a parameter indicating whether
or not the entire version is also being deleted. This preserves the
current "optimization" that if we are deleting the entire version, we
don't "need" to re-run the concept transforms.

CC-2651