RVA-24 When changing a version slug, delete old access points
If you change the slug of an existing version (which a Portal user can
do by changing the version's title), and that version has import
and/or publish flags set, and you use the force workflow flag, you get
new access points generated that use the new version slug.

But any existing access points using the old version slug aren't
deleted. To be specific: the database rows for the old version slug
are indeed made historical, but the Sesame repo and/or SISSVoc spec
file still exist.

In this situation, we need to run appropriate delete subtasks for the
old access points as well as the insert subtasks for the new ones.

(As noted above: this is about doing delete subtasks for all access
points that use the version slug in some way. For now, that means
access points related to import (apiSparql and sesameDownload) and
publish (sissvoc).)

In fact, let's take the opportunity to fix this properly: replace the
access points even if the force-workflow flag is _not_ set.

In fact, let's do even better: trigger the replacement also if the
_owner_ or _vocabulary slug_ is changed. This gets us well on the way
to supporting arbitrary metadata changes.

CC-2661