RVA-24 Trigger renewing access points when only owner changes
Commit 926ff92635b7c9870a991e9a747dbe434cca9fe4 added a "slugChanged"
flag to the logic of VersionsModel, and the flag is set also if only
the vocabulary owner has changed.

But the code to _use_ the flag is nested in a conditional for which
the test wasn't satisfied if only the vocabulary owner has changed.

So add slugChanged as part of the test.

Take the opportunity to fix the comment preceding the test to make it
match the test.

CC-2661