During test, after clearing the Owners cache, put back "*" key
I noticed that the JaCoCo report showed for the class
OwnersCacheLoader that this line of code:

  // Hmm, somehow the key was ejected from the cache.
  // Can't happen?
  return Owners.ALL_OWNERS_OWNER_ID;

was indeed getting coverage! Tracked this down to the clearing of the
cache that's done between tests.

In the test helper method Owners.clear(), after clearing the cache,
put back the "magic" mapping "*" -> 0 so that the cache looks the way
it does after Owners.init().