Add test of changing the do-import flag
Add a test testApplyChangesCurrentVoVREVeAPVAW11 which exercises the
workflow in the case that the do-import flag is changed for
versions. We enable the do-import flag for each of three versions, two
of which are superseded. We check that the Sesame importer import
operation is run for each version. We then disable the flag for each
version, and check that the Sesame importer delete operation is run
for each version. The ResourceMap transform delete operation is also
run in this case.