CC-2704 Add test for defect with re-importing on PP harvest delete
Consider a vocabulary that is based on a PoolParty project, and that
has an existing version such that:
* do-poolparty-harvest="true"
* do-import="true"
* the version has a file access point

Now update the version, but set do-poolparty-harvest="false".

In this case, a Sesame import subtask must be run (together with the
usual follow-up SesameInsertMetadata transform), to clear out the
concept data that was previously harvested, and to re-import just the
concept data from the file access point.  But that is not
happening. No Sesame import subtask (with SesameInsertMetadata
transform) is run.

Add a test that shows this defective behaviour. For now, disable the
test.