CC-2704 Adjust test data for testApplyChangesCurrentVoVREVeAPVAW12
See the comment of the previous
commit (ea4b0abfdf9ed99ad2544921c94ca2215a338466).

That comment said, "(together with the usual follow-up
SesameInsertMetadata transform)". But that was wrong. We only use
the SesameInsertMetadata transform when importing data from
PoolParty.

Adjust the test data so as not to expect the SesameInsertMetadata
transform to be run.