Rename test helper method compareJson as compareJsonFiles
In future commits I'm going to introduce more methods for comparing
actual and expected JSON data. The test helper method compareJson()
compares two files containing JSON, and it takes two parameters that
are both filenames.

Rename the method as "compareJsonFiles" to make this (more) explicit,
as the existing name "compareJson" isn't as clear as it could be.