CC-2735 Fix deletion of user-specified access points
The work for CC-2651 (RVA-23) introduced a regression: you can't
delete a user-specified access point of type apiSparql,
sesameDownload, or sissvoc. The Registry log file shows a
NullPointerException at
AccessPointsModel$UpdateCurrentVisitor.visitDeleteCommand(), line
866.

The method WorkflowMethods.deleteAccessPoint() had a return type
that was a (possibly-empty) list of subtasks to be run. The
previous CC changed its return type to a pair of values, of which
the second is the list of subtasks, and the first is a boolean
value indicating whether the method's caller should delete (i.e.,
make historical) the database row for the access point.

The code change updated some of the cases of the switch statement
in the method, but missed the situations of a user-specified
access point of type apiSparql, sesameDownload, or sissvoc. The
result: the value returned in those cases was still
null (representing an empty-list of subtasks) instead of the
now-expected pair of values. The return value should be as it now
is for access points of type webPage, i.e., Pair.of(true, null).

The caller,
AccessPointsMode$UpdateCurrentVisitor.visitDeleteCommand(), line
866, expects the return value to be non-null, and therefore
generates a NullPointerException.

I changed WorkflowMethods.deleteAccessPoint() to return the
correct value, instead of null. I added test cases to confirm the
correct behaviour.

NB: although the new test cases include deleting a user-specified
sesameDownload access point, users are not supposed to be able to
create such an access point – and indeed, they can't. (See
CC-2406.) Nevertheless, it's necessary for the
WorkflowMethods.deleteAccessPoint() method to support deleting
such an access point.

RVA-87