CC-2725 Add /adminApi/solr/commit method to force soft commit
On publishing a vocabulary, there's a step at the end to index the
vocabulary in Solr.

We currently configure the two Solr collections with an autoSoftCommit
maxTime of 10 seconds.

That means, you may have to wait up to 10 seconds before you see the
vocabulary and its concepts in searches.

That's fine in normal use, but not very convenient during automated
testing.

Add an admin API method /adminApi/solr/commit to force a soft commit.

RVA-82