CC-2709 Compute resource sissvoc_endpoint only when indexing
The creation of the Solr documents for resource data is done in two
stages:

* There's a ResourceDocs workflow transform that creates a version
  artefact (of type RESOURCE_DOCS) that is a file containing a JSON
  array of Solr documents, one document for each resource in a
  version, each document containing most of the fields that are to go
  into it.

* During the vocabulary indexing process (which happens after all of
  the workflow's subtasks are completed), the version artefacts for
  all versions of a vocabulary are read in, and each Solr document is
  supplemented with fields for the current top-level and version
  metadata, e.g., vocabulary and version titles, IDs, last_updated,
  subject labels, publishers, etc., and then sent off to Solr.

The point of this division of labour is to avoid putting into the
version artefacts any vocabulary/version metadata that could change,
and only accessing that metadata "just in time", i.e., just before
sending the Solr documents to Solr. This means, for example, that we
don't have to re-process all of the concept data for every version in
a vocabulary, when a new version is added to the vocabulary (which
changes the last_updated value for every resource in the vocabulary),
or if a publisher is added.

There's a defect in the division of labour: the ResourceDocs workflow
transform inserts the field sissvoc_endpoint, which the Portal uses to
generate the "View resource as linked data" links in resource search
results. But SISSVoc endpoints contain slugs!

The ResourceDocs workflow transform isn't run if only
vocabulary/version metadata changes. (And it is correct that it isn't
run.) In particular, it isn't run if only the version slug and title
change (which will be the case if a Portal CMS user changes the
version title).

The result: if you change the version slug, then resource search
results for resources in that version break: the "View resource as
linked data" links still point to the old SISSVoc endpoint.

Move the insertion of the sissvoc_endpoint field from the ResourceDocs
workflow transform to the indexing
process (EntityIndexer.indexResourceDocsForVocabulary()).

(It turns out that this defect also broke the special case of rehoming
a server, where we use the build task migrate-accesspoint-hostname to
change base URLs of access points in the Registry database. Because of
this defect, merely re-indexing after running the build task isn't
enough.)

RVA-89