Fix "skos:definition:" for testJsonTreeTransformProvider1 test data
The test data for testJsonTreeTransformProvider1 includes occurrences
of "skos:definition:" with a bogus trailing colon. Remove the colon
from the input data, and adjust the expected results to include the
corresponding definitions.

NB: the data for the tests testConceptTreeTransformProvider{1,2} and
testResourceDocsTransformProvider1 still includes instances of the
same thing; they will be fixed in a later commit.