CC-2763 Browse: make new concept ResourceRefs if sort-by-notation
The old browse implementation had a defect which we didn't already
fix.

If the maySortByNotation flag is set, and there's a polyhierarchy, and
a concept's ResourceRef singleton instance is used in more than one
place, the generated notationSortOrder values could be wrong.

If the maySortByNotation flag is set, then singleton instances are
wrong; when making a concept_ref, always make a new ResourceRef
instance.

RVA-107