CC-2522 ConceptTree: fix children of concepts not in a CS
If includeConceptSchemes is true, then we still show concepts that are
not members of any concept scheme; they appear at the top level,
arranged according to the broader/narrower hierarchy.  The concepts
shown at the top level are the correct ones. But there was a defect in
the navigation from the top level down: you could encounter concepts
that are members of concept schemes.

Now, make sure that if you navigate down from a concept at the top
level, you only encounter concepts not in any concept scheme.

RVA-46