CC-2751 Use millisecond precision for start_date/end_date db columns
The database's main tables, such as vocabularies, versions, and
access_points, have columns start_date and end_date, that indicate the
temporal validity of each row.

For example, if a row of the versions table has start_date =
'2016-09-27 04:23:41' and end_date = '2019-10-09 01:25:09', then that
contents of that row represented "the" current instance of that
version between those two dates/times (including the very instant of
the start date/time, but excluding the very instant of the end
date/time).

It is extremely important for the consistency of the database that
these start_date and end_date values be correct and consistent. In
particular, it must always be the case, for every row, that start_date
<= end_date. We don't have every single possible protection to prevent
the persisting of a row to a table for which start_date > end_date,
but we do have a VersionListener class that is a type of "gatekeeper"
that checks all attempts to persist (i.e., either insert or update) a
row to the versions table: it checks that start_date <= end_date, and
forbids the insert/update if that test fails. In practice, we've never
seen this check fail ... until now.

Now note that the two example dates/times given above have second
precision, i.e., the times '04:23:41' and '01:25:09' indicate the
hour, minute, and second, but not, for example, milliseconds. And
second precision is indeed the precision used by all start_date and
end_date values in the database, because that happens to be MySQL's
default precision of columns that have been created using the type
DATETIME.

But Java's LocalDateTime values, as used internally by the Registry,
represent dates/times to nanosecond precision. It's slightly more
complicated than that: when the Registry code fetches the current
time (using Clock.systemUTC()), in Java 8 it gets back a value with
millisecond precision. So not all of the available precision of the
LocalDateTime type is used by the Registry. The resolution is
apparently increased in Java 9, but I haven't yet checked that out.

So what happens if an API user does two updates within the same
second?

For example, consider what happens if we do an update to a vocabulary,
the result of which is to insert a row into the versions table that
has start_date = '2020-07-22 06:03:19.625'. Here, we show millisecond
precision, as used by the Registry.

Now, suppose we delete that vocabulary at '2020-07-22 06:03:19.825',
i.e., within the same second.

The Registry will load the row that was only just previously inserted
into the versions table, and set its end_date to '2020-07-22
06:03:19.825', and then that row will be persisted to the database.

It turns out that this fails.

When the row was inserted into the database, because the column type
is DATETIME, MySQL rounded the date/time value to the nearest second!

So when the Registry received the request to delete the vocabulary, it
loaded that row back from the database, and it got a value for
start_date that was '2020-07-22 06:03:20', i.e., with the date/time
rounded 0.375 second into the future!

Then, after setting end_date to '2020-07-22 06:03:19.825' and
requesting that the row be persisted, the JPA persistence layer
invoked the VersionListener, which "correctly" determined that the
row's start_date value was after its end_date value, and proceeded to
throw an exception, which caused the entire delete request to fail,
albeit gracefully, and with no damage to the database integrity, but
with an "internal server error" sent back to the caller.

We never noticed this before, because we don't have API users who
insert a version row and then do an update of it within the same
second, but Joel's automated tests now include a test in which exactly
this scenario happens.

The solution is to alter the definition of every start_date and
end_date column, to change the column type from DATETIME to
DATETIME(6), so that every value will now be stored with microsecond
precision. (Here, "6" means six decimal digits of precision for
seconds, which is as many as MySQL supports. Sigh: it seems that in
the SQL standard, the default precision of seconds within date/time
columns is six digits, and up to nine digits are supported, but this
is yet another case where MySQL deviates from the SQL standard.) This
change can be effected easily enough using the Registry's nice support
for database schema migration.

A workaround for automated testing is to insert a one-second sleep
between the API method invocations that first insert the version row
and then update that row (in this particular case, which is being done
as part of deleting the whole vocabulary).

Side note #1: MySQL's behaviour changed from version 5.5 to 5.6. In
5.5, fractional second values were truncated, not rounded. See
https://bugs.mysql.com/bug.php?id=68760 .

Side note #2: but surely the Registry's own test suite should already
be triggering this problem? Ah, but the Registry's test suite uses an
embedded H2 database – not MySQL – and H2 implements the SQL standard:
its TIMESTAMP type by default has six decimal digits of precision for
seconds: more than enough to represent the millisecond-precision
values created by the Registry.

RVA-146