CC-2522 ConceptTree: tidy ResourceOrRef.compareTo()
In one of the early iterations of the implementation of
ResourceOrRef.compareTo(), we needed to do a cast to get access to
some methods implemented only in a subclass. We subsequently
refactored the methods so that the cast was no longer necessary, but
didn't tidy up the compareTo() method.

Tidy the coding of ResourceOrRef.compareTo().