CC-2522 ConceptTree: when breaking a cycle, remove any sort order
If after DFS is	completed, there are non-visited nodes,	we have a
cycle.  We choose one to be a new root, and continue the DFS.

But the	set of roots is	a TreeSet, so all its elements must be
comparable.

It's possible for the chosen non-visited node to be a collection that
is itself a member of an ordered collection, and therefore it's
possible for that node to have had an orderedCollectionSortOrder value
assigned. In that case, compareTo() will fail with an NPE when
compared with any member of roots that itself did not have an
orderedCollectionSortOrder value assigned.

Since we won't be producing a final result in this case, it's OK for
us to reset the chosen node's orderedCollectionSortOrder value to
null. The compareTo() method will then not fail.